# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Crusher API
  description: |
    碎石记（Crusher）API 文档

    ## 认证
    大多数端点需要 Bearer Token 认证。Token 通过 Supabase Auth 获取。

    ## 错误响应
    所有错误响应遵循统一格式：
    ```json
    {
      "success": false,
      "error": "错误消息",
      "code": "错误码（可选）",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@crusher.com
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://crusher.com
    description: 生产环境
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: '错误消息'
        code:
          type: string
          example: 'ERROR_CODE'
        details:
          type: object
    ApiMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true

    # 用户相关
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        credits:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # 文档相关
...
paths:
...
  /api/admin/users:
    get:
      tags:
        - Admin
      summary: 获取用户列表（管理员）
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功返回用户列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Profile'
                      meta:
                        $ref: '#/components/schemas/ApiMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  responses:
    Unauthorized:
      description: 未授权（未登录或 Token 无效）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Unauthorized'
            code: 'UNAUTHORIZED'
    Forbidden:
      description: 禁止访问（权限不足）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Forbidden'
            code: 'FORBIDDEN'
...